moves rest of constants, types to the core models directory
