--- src/api/zero-x-v2/constants.ts
@@ -1,4 +1,0 @@
-export const ZERO_X_V2_BASE_URL = 'https://api.0x.org/';
-export const ZERO_X_EXCHANGE_ALLOWANCE_HOLDER_ADDRESS = '0x0000000000001fF3684f28c67538d4D072C22734';
-export const ZERO_X_PROXY_BASE_TOKEN_ADDRESS =
-  '0xeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeee';
