+++ src/api/zero-x-v2/index.ts
@@ -1,4 +1,2 @@
 export * from './zero-x-v2-quote';
-export * from './types';
-export * from './constants';
 export * from './errors';
